refactor: widgets 분리
- CouponManager, ProductManager
