+++ src/refactoring/widgets/index.ts
@@ -1,2 +1,4 @@
 export { ProductSection } from './ProductSection';
 export { CartSection } from './CartSection';
+export { ProductManager } from './ProductManager';
+export { CouponManager } from './CouponManager';
